lkl tools: Use tools/build as build system.
This patch introduces a build system to tools/lkl that is more in line
with the kernel build system Kbuild. It tries to retain most of the
behaviour of the original Makefile with one major difference: the clean
target only cleans the tools/lkl tree and not the whole kernel tree.
I think this is more appropriate, also considering the following patches.

Some small parts of the new Makefile are stolen from tools/perf/Makefile.perf.

Signed-off-by: Thomas Liebetraut <thomas@tommie-lie.de>